Extract getSelectedNode into a shared lexical-utils module
